etl/requirements.txt
python-dotenv==1.0.0
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg2==2.9.9
sqlalchemy-bigquery==1.11.0